Added check for nvim 0.5
This logic should pick up the 0.5 nightly version of nvim that supports `startinsert` command. Else it will revert for feedkeys for all other neovim versions below 0.5. We may need to revise in the future versions of nvim. But at this point that looks some time away.

Tested on 0.5 and works for me. Will need testing on other versions of nvim/vim to confirm all is well.